requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.2
django-crispy-forms==2.0
djangorestframework==3.14.0
Pillow==9.5.0
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
